src/main/java/com/easy/store/backend/context/category/domain/model/Category.java
package com.easy.store.backend.context.category.domain.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.sql.Timestamp;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class Category {
    private Long id;
    private String name;
    private String description;
    private String imageName;
    private Long createBy;
    private Long updateBy;
    private Timestamp creationDate;
    private Timestamp updateDate;
    private String state;

    public boolean isValid(Category category) {
        if(category.getName() == null || category.getDescription() == null) return false;

        return !category.getName().isEmpty() && !category.getDescription().isEmpty();
    }
}
